Factor ConfigContext into an interface and allow for a supplier for test configs
