+++ Configuration/src/main/java/io/deephaven/configuration/ParsedProperties.java
@@ -3,16 +3,16 @@
 //
 package io.deephaven.configuration;
 
+import io.deephaven.internal.log.LoggerFactory;
+import io.deephaven.io.logger.Logger;
+import org.apache.commons.lang3.StringEscapeUtils;
+import org.jetbrains.annotations.NotNull;
+
 import java.io.*;
 import java.util.*;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
-
-import io.deephaven.io.logger.Logger;
-import io.deephaven.internal.log.LoggerFactory;
-import org.apache.commons.lang3.StringEscapeUtils;
-import org.jetbrains.annotations.NotNull;
 
 /**
  * Class for reading in a customized properties file, applying only the locally-relevant properties and keeping track of
@@ -95,7 +95,7 @@ public class ParsedProperties extends Properties {
     @SuppressWarnings("WeakerAccess")
     // This is used from an external utility for showing useful information about prop files.
     public ParsedProperties() {
-        this(false);
+        this(false, new DefaultConfigurationContext());
     }
 
     /**
@@ -104,8 +104,8 @@ public class ParsedProperties extends Properties {
      * @param ignoreScopes True if this parser should ignore scope restrictions, false otherwise. Used by the
      *        PropertyInspector when checking whether required or disallowed properties are present.
      */
-    public ParsedProperties(final boolean ignoreScopes) {
-        context = new ConfigurationContext();
+    public ParsedProperties(final boolean ignoreScopes, final @NotNull ConfigurationContext context) {
+        this.context = context;
         finalProperties = new HashSet<>();
         lineNumbers = new HashMap<>();
         props = new LinkedHashMap<>();
